feat: implement dragging to center notch & wallpaper selector, weather service

+++ src/lib/theme/matugen.ts
@@ -23,16 +23,24 @@ export const callMatugen = async (
 
 export const relinkTheme = async (themeDir: string): Promise<void> => {
   await Promise.all(
-    Object.entries(THEME_SYMLINKS).map(async ([file, { dest, reload }]) => {
-      const src = buildPath(themeDir, file);
-      const destFile = Gio.File.new_for_path(dest);
+    Object.entries(THEME_SYMLINKS).map(
+      async ([file, { dest, reload, copy }]) => {
+        const src = buildPath(themeDir, file);
+        const destFile = Gio.File.new_for_path(dest);
 
-      // biome-ignore lint/style/noNonNullAssertion: <get_parent() and get_path() are non-null for absolute paths>
-      ensureDir(destFile.get_parent()!.get_path()!);
+        // biome-ignore lint/style/noNonNullAssertion: <get_parent() and get_path() are non-null for absolute paths>
+        ensureDir(destFile.get_parent()!.get_path()!);
 
-      if (destFile.query_exists(null)) destFile.delete(null);
-      destFile.make_symbolic_link(src, null);
-      if (reload) await exec(reload).catch(() => {});
-    }),
+        if (copy) {
+          const srcFile = Gio.File.new_for_path(src);
+          srcFile.copy(destFile, Gio.FileCopyFlags.OVERWRITE, null, null);
+        } else {
+          if (destFile.query_exists(null)) destFile.delete(null);
+          destFile.make_symbolic_link(src, null);
+        }
+
+        if (reload) await exec(reload).catch(() => {});
+      },
+    ),
   );
 };
